Fix nilpointers in spaces shares, clean up conversions
